requirements.txt
pennylane~=0.43.1
pennylane-calculquebec
python-dotenv
numpy
pygame
matplotlib
